move benchmark thread pool to own file
