+++ bench/benchmark_dumps.py
@@ -1,13 +1,12 @@
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
 from json import loads as json_loads
-from queue import Queue
-from threading import Thread
 
 import pytest
 
 import orjson
 
 from .data import fixtures, libraries
+from .simple_thread_pool import SimpleThreadPool
 from .util import read_fixture, read_fixture_obj
 
 
@@ -20,42 +19,6 @@ def test_dumps(benchmark, fixture, library):
     data = read_fixture_obj(f"{fixture}.xz")
     benchmark.extra_info["correct"] = json_loads(dumper(data)) == data
     benchmark(dumper, data)
-
-
-class SimpleThreadPool:
-    class Worker(Thread):
-        def __init__(self, tasks: Queue):
-            Thread.__init__(self, daemon=True)
-            self.tasks = tasks
-            self.start()
-
-        def run(self):
-            while True:
-                task = self.tasks.get()
-                try:
-                    if task is None:
-                        break
-                    func, args, kwargs = task
-                    func(*args, **kwargs)
-                finally:
-                    self.tasks.task_done()
-
-    def __init__(self, num_threads: int):
-        self.tasks: Queue = Queue()
-        self.workers = [SimpleThreadPool.Worker(self.tasks) for _ in range(num_threads)]
-
-    def add_task(self, func, *args, **kwargs):
-        self.tasks.put((func, args, kwargs))
-
-    def wait_completion(self):
-        self.tasks.join()
-
-    def close(self):
-        self.tasks.join()
-        for _ in self.workers:
-            self.tasks.put(None)
-        for worker in self.workers:
-            worker.join()
 
 
 @pytest.mark.parametrize("threads", [1, 2, 3, 4])
